put crash report utils methods in file utils instead
